Install the typing package unconditionally
The typing package is part of the stdlib for Python >=3.5, so installing the typing
package is not necessary on these versions. However, restricting its installation
to Python versions < 3.5 will prevent the conda package from being "noarch". Since
the stdlib takes precedence over installed packages, having an unnecessary typing
package should not hurt. This should be removed when support for Python < 3.5 is
no longer an issue.